+++ src/test/java/com/iexec/worker/config/SchedulerConfigurationTests.java
@@ -28,8 +28,13 @@ class SchedulerConfigurationTests {
 
     @Test
     void shouldCreateBeanInstance() {
-        runner.withPropertyValues("core.protocol=http", "core.host=localhost", "core.port=13000")
+        runner.withPropertyValues("core.protocol=http", "core.host=localhost", "core.port=13000", "core.poolAddress=0x365E7BABAa85eC61Dffe5b520763062e6C29dA27")
                 .withConfiguration(UserConfigurations.of(SchedulerConfiguration.class))
-                .run(context -> assertThat(context).hasSingleBean(SchedulerClient.class));
+                .run(context -> {
+                    assertThat(context).hasSingleBean(SchedulerClient.class);
+                    assertThat(context).getBean("schedulerConfiguration", SchedulerConfiguration.class)
+                            .extracting("poolAddress")
+                            .isEqualTo("0x365E7BABAa85eC61Dffe5b520763062e6C29dA27");
+                });
     }
 }